+++ src/test/java/com/thebank/loan/entity/RiskAssessmentEntity.java
@@ -1,39 +1,38 @@
 package com.thebank.loan.entity;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 public class RiskAssessmentEntity {
-    private BigDecimal probabilityOfPayment; // entre 0 e 1
-    private BigDecimal debtToIncomeRatio;
-    private BigDecimal zipcodeRiskFactor;
-    private BigDecimal finalRiskScore; // 0 = baixo risco, 100 = alto risco
+    private Double probabilityOfPayment; // entre 0 e 1
+    private Double debtToIncomeRatio;
+    private Double zipcodeRiskFactor;
+    private Double finalRiskScore; // 0 = baixo risco, 100 = alto risco
 
-    public BigDecimal getProbabilityOfPayment() {
+    public Double getProbabilityOfPayment() {
         return probabilityOfPayment;
     }
-    public RiskAssessmentEntity setProbabilityOfPayment(BigDecimal probabilityOfPayment) {
+    public RiskAssessmentEntity setProbabilityOfPayment(Double probabilityOfPayment) {
         this.probabilityOfPayment = probabilityOfPayment;
         return this;
     }
-    public BigDecimal getDebtToIncomeRatio() {
+    public Double getDebtToIncomeRatio() {
         return debtToIncomeRatio;
     }
-    public RiskAssessmentEntity setDebtToIncomeRatio(BigDecimal debtToIncomeRatio) {
+    public RiskAssessmentEntity setDebtToIncomeRatio(Double debtToIncomeRatio) {
         this.debtToIncomeRatio = debtToIncomeRatio;
         return this;
     }
-    public BigDecimal getZipcodeRiskFactor() {
+    public Double getZipcodeRiskFactor() {
         return zipcodeRiskFactor;
     }
-    public RiskAssessmentEntity setZipcodeRiskFactor(BigDecimal zipcodeRiskFactor) {
+    public RiskAssessmentEntity setZipcodeRiskFactor(Double zipcodeRiskFactor) {
         this.zipcodeRiskFactor = zipcodeRiskFactor;
         return this;
     }
-    public BigDecimal getFinalRiskScore() {
+    public Double getFinalRiskScore() {
         return finalRiskScore;
     }
-    public RiskAssessmentEntity setFinalRiskScore(BigDecimal finalRiskScore) {
+    public RiskAssessmentEntity setFinalRiskScore(Double finalRiskScore) {
         this.finalRiskScore = finalRiskScore;
         return this;
     }
